fix: data.sql 수정 (#noissue)

+++ src/main/java/com/kcc/pms/domain/feature/controller/FeatureController.java
@@ -26,7 +26,7 @@ public class FeatureController {
 
     @GetMapping("/register")
     public String featureInfo(HttpSession session){
-        Integer prjNo = (Integer) session.getAttribute("prjNo");
+        Long prjNo = (Long) session.getAttribute("prjNo");
         System.out.println("prjNo = " + prjNo);
         return "feature/featureInfo";
     }
